user microservice/requirements.txt
Django==4.2.6
confluent-kafka==2.3.0
djangorestframework==3.14.0
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.0
django-filter==23.3
requests==2.31.0 
psycopg2-binary==2.9.2 
gunicorn==20.1.0